MA-10031 [Android]fix compile warning for android
There are some warning on 64-bit platform,like:
drivers/usb/gadget/bcb.c:151:29: warning: cast from pointer to integer
of different size [-Wpointer-to-int-cast]
sprintf(addr_str, "0x%x", (unsigned int)p_block);

fix through cast point to uintptr_t firstly.

declare in header to fix below warning:
drivers/usb/gadget/f_fastboot.c:971:5: warning: implicit declaration of
function 'fastboot_tx_write_str' [-Wimplicit-function-declaration]
fastboot_tx_write_str("FAILinvalid mmc device");

+++ drivers/usb/gadget/bcb.c
@@ -148,7 +148,7 @@ int bcb_rw_block(bool bread, char **ppblock,
 			(int)(blk_size * block_cnt));
 			return -1;
 		}
-		sprintf(addr_str, "0x%x", (unsigned int)p_block);
+		sprintf(addr_str, "0x%x", (unsigned int)(uintptr_t)p_block);
 		ret = do_raw_read(NULL, 0, 6, argv);
 		if (ret) {
 			free(p_block);
@@ -159,7 +159,7 @@ int bcb_rw_block(bool bread, char **ppblock,
 		*ppblock = p_block;
 		*pblksize = (uint)blk_size;
 	} else {
-		sprintf(addr_str, "0x%x", (unsigned int)pblock_write);
+		sprintf(addr_str, "0x%x", (unsigned int)(uintptr_t)pblock_write);
 		ret = do_write(NULL, 0, 6, argv);
 		if (ret) {
 			printf("do_write failed, ret %d\n", ret);
